0.0.2

+++ TMS/dbo/Tables/Lines.sql
@@ -1,6 +1,6 @@
 CREATE TABLE [dbo].[Lines]
 (
-  [ID] INT NOT NULL PRIMARY KEY,
+  [ID] BIGINT NOT NULL PRIMARY KEY,
   [latitude] DECIMAL(9,7) NOT NULL,
   [longitude] DECIMAL(8,7) NOT NULL,
   [LineCode] CHAR(3) NOT NULL,
